refactor(order): tighten get_my_orders surface and pin its branches (DEFI-2826)
Address self-review on the get_my_orders query:

- Return Result<_, GetMyOrdersError> from the lib dispatcher so the
  malformed-cursor path is unit-testable; the public query still traps on
  it (a bad opaque cursor is a client bug). Drops the in-dispatcher panic.
- Narrow length and MAX_ORDERS_PER_RESPONSE to u32 (matching MAX_DEPTH_LIMIT
  / MAX_FILTER_LEN) and remove the unreachable usize::try_from expect; the
  over-cap length keeps clamping silently as documented.
- Add dispatcher unit tests: malformed cursor, length cap, length 0.
- Tighten the integration timestamp assertion to a [before, after] window.
- Bench: stop paginating once the known total is reached, shaving the extra
  empty page call when the total is an exact multiple of the page size.
- Document that UserOrder.order.owner is always the caller.

+++ libs/types/src/lib.rs
@@ -221,7 +221,7 @@ pub struct OrderRecord {
 
 /// Maximum number of orders returned by a single [`get_my_orders`] call.
 /// Requests for more are silently capped to this many.
-pub const MAX_ORDERS_PER_RESPONSE: u64 = 100;
+pub const MAX_ORDERS_PER_RESPONSE: u32 = 100;
 
 /// Request for the `get_my_orders` query: a page over the caller's orders,
 /// newest first. `length` is capped at [`MAX_ORDERS_PER_RESPONSE`].
@@ -234,7 +234,7 @@ pub struct GetMyOrdersArgs {
     /// `None` starts from the newest order.
     pub after: Option<OrderId>,
     /// Maximum number of orders to return.
-    pub length: u64,
+    pub length: u32,
 }
 
 /// One entry in a [`get_my_orders`] response: an order the caller placed.
@@ -244,7 +244,9 @@ pub struct UserOrder {
     pub id: OrderId,
     /// The trading pair the order was placed on.
     pub pair: TradingPair,
-    /// The full order record.
+    /// The full order record. `get_my_orders` only returns the caller's own
+    /// orders, so `order.owner` is always the caller — reused as-is for shape
+    /// parity with other order-returning endpoints.
     pub order: OrderRecord,
 }
